refactor: 생성 메서드명 of로 통일, vo 불변객체로 수정, UUID 생성방법 수정
